Extract GraphQL access
